pyproject.toml
[project]
name = "node-editor-rf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml", 
    "openai", 
    "pillow", 
    "python-dotenv", 
    "requests",
    "torch",
    "diffusers",
    "transformers",
    "accelerate",
	"xformers"
]